esummit-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0
gunicorn==21.2.0
WTForms==3.1.1
Werkzeug==3.0.1
celery==5.3.6
redis==5.0.1
